Allowed decimals on dashboard

+++ mqtt/helper.py
@@ -35,17 +35,17 @@ def connect_mqtt(client_id=client_id) -> mqtt_client:
 def store(msg):
     channel_layer = get_channel_layer()
     #update associated model
-    topics[msg.topic]['model'].objects.create(date=datetime.now(), data=int(msg.payload.decode()), trip=Trip.objects.last()) 
+    topics[msg.topic]['model'].objects.create(date=datetime.now(), data=float(msg.payload.decode()), trip=Trip.objects.last()) 
 
     if str(msg.topic) != "/DAQ/Latitude" and str(msg.topic) != "/DAQ/Longitude":
         #do NOT deal with long/lat here. need to implement separate feature to store it
         pass
     else:
         #send to team view always, except for lat/long data
-        async_to_sync(channel_layer.group_send)("teamdata", {"type": f"team.notif", "module": f"{topics[msg.topic]['name']}", "content": int(msg.payload.decode()), "error": False})
+        async_to_sync(channel_layer.group_send)("teamdata", {"type": f"team.notif", "module": f"{topics[msg.topic]['name']}", "content": float(msg.payload.decode()), "error": False})
     if str(msg.topic) in ['/DAQ/Speed', "/Power_Control/Voltage", "/Power_Control/Current"]:
         #send to dashboard ONLY for speed and energy(to avoid sending non-relevant data)
-        async_to_sync(channel_layer.group_send)("speed", {"type": f"data.notif", "module": f"{topics[msg.topic]['name']}", "content": int(msg.payload.decode()), "error": False})
+        async_to_sync(channel_layer.group_send)("speed", {"type": f"data.notif", "module": f"{topics[msg.topic]['name']}", "content": float(msg.payload.decode()), "error": False})
 
 
 def subscribe(topic, client: mqtt_client):
